move hook to separate file
